Add option -p
If -p is given the files linked to the output dir are limited to source
files that have the path given with -p in the name.

This is usefull if you have a couple of mail accounts, and want to
search only one account.

Example:

I have three email accounts I sync with offlineimap and index them all
together with one notmuch index.
To search globally, I run:
$ mutt-notmuch -g ~/mail/temp/search
To search only my work email, I run:
$ mutt-notmuch -g -p 'mail/work' ~/mail/temp/search

I know I could have done this with multiple notmuch databases but that
was too much hassle with environment variables, especially with the
combination of offlineimap, mutt, notmuch etc